requirements.txt
Flask==1.1.2
cassandra-driver==3.24.0